small changes to resolve git comments
