+++ components/aws/sagemaker/tests/unit_tests/.gitignore
@@ -1,0 +1,2 @@
+*.log
+.coverage
