+++ components/aws/sagemaker/tests/unit_tests/requirements.txt
@@ -1,0 +1,5 @@
+boto3
+coverage
+pytest
+pyyaml
+sagemaker
