+++ components/aws/sagemaker/tests/unit_tests/run_all_tests.sh
@@ -1,0 +1,4 @@
+export PYTHONPATH=../../
+
+coverage run -m pytest --ignore=tests/test_utils.py --junitxml ./unit_tests.log
+coverage report --omit "*/usr/*,tests/*,*__init__*,*/Python/*"
