+++ components/aws/sagemaker/tests/unit_tests/tests/test_deploy.py
@@ -5,29 +5,28 @@ from unittest.mock import patch, Mock, MagicMock
 from botocore.exceptions import ClientError
 from datetime import datetime
 
-from workteam.src import workteam
+from deploy.src import deploy
 from common import _utils
-from . import testUtils
+from . import test_utils
 
 
 required_args = [
   '--region', 'us-west-2',
-  '--team_name', 'test-team',
-  '--description', 'fake team'
+  '--model_name_1', 'model-test'
 ]
 
-class workteamTestCase(unittest.TestCase):
+class DeployTestCase(unittest.TestCase):
   @classmethod
   def setUpClass(cls):
-    parser = workteam.create_parser()
+    parser = deploy.create_parser()
     cls.parser = parser
 
   def test_sample(self):
     args = self.parser.parse_args(required_args)
-    response = _utils.create_workteam_request(vars(args))
-    self.assertEqual(response['WorkteamName'], 'test-team')
+    response = _utils.create_endpoint_config_request(vars(args))
+    self.assertEqual(response['EndpointConfigName'], 'EndpointConfig-test')
 
   def test_empty_string(self):
     args = self.parser.parse_args(required_args)
-    response = _utils.create_workteam_request(vars(args))
-    testUtils.check_empty_string_values(response)
+    response = _utils.create_endpoint_config_request(vars(args))
+    test_utils.check_empty_string_values(response)
