+++ components/aws/sagemaker/tests/unit_tests/tests/test_ground_truth.py
@@ -7,7 +7,7 @@ from datetime import datetime
 
 from ground_truth.src import ground_truth
 from common import _utils
-from . import testUtils
+from . import test_utils
 
 
 required_args = [
@@ -39,4 +39,4 @@ class GroundTruthTestCase(unittest.TestCase):
   def test_empty_string(self):
     args = self.parser.parse_args(required_args)
     response = _utils.create_labeling_job_request(vars(args))
-    testUtils.check_empty_string_values(response)
+    test_utils.check_empty_string_values(response)
