--- components/aws/sagemaker/unit_tests/requirements.txt
@@ -1,3 +1,0 @@
-boto3
-botocore
-sagemaker
