--- components/aws/sagemaker/unit_tests/run_all_tests.sh
@@ -1,4 +1,0 @@
-export PYTHONPATH=../
-
-coverage run -m pytest --junitxml ./log
-coverage report --omit "*/usr/*,tests/*,*__init__*,*/Python/*"
